Remove notification client's k8s dependency and move to internal
